requirements.txt
streamlit
pdfplumber
python-docx
pandas
scikit-learn
sentence-transformers
openpyxl
plotly
openai
Pillow
# torch
